+++ lib/presentation/main/user/main_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:root/components/widgets/root_layout.dart';
-import 'package:root/presentation/main/detail_main_screen.dart';
+import 'package:root/presentation/main/user/detail_main_screen.dart';
 import 'package:root/components/widgets/root_main_app_bar.dart';
 
-import '../../components/root_colors.dart';
+import '../../../components/root_colors.dart';
 
 class MainScreen extends StatefulWidget {
   const MainScreen({super.key});
